bug(PMD): fix PMD god class flag
- Remove reader async task into its own package

- Fix the produced bug storm

[Delivers:

164046972]
