+++ app/src/main/res/layout/nfc_security_dashboard.xml
@@ -14,10 +14,9 @@
             android:id="@+id/asset_details_progress_bar"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            app:layout_constraintTop_toBottomOf="@id/addSerial"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
-            android:layout_marginTop="80dp"
+            app:layout_constraintTop_toBottomOf="@id/scanNfcTitleTextView"
             >
 
             <ProgressBar
